+++ package.json
@@ -8,7 +8,7 @@
     "@testing-library/user-event": "^12.8.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
