+++ src/App.css
@@ -1,90 +1,149 @@
-body{
-    background-image: url('./assets/images/background-body.jpeg');
-    font-family: 'Roboto', sans-serif;
+@import url('https://fonts.googleapis.com/css2?family=Oswald:wght@300&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap');
+
+* {
+    margin: 0;
+    padding: 0;
+    font-family: Oswald;
+    box-sizing: border-box;
 }
 
-h1{
-color:#763D09;
-font-weight: 400;
-font-size: 35px;
-font-family: 'Oswald', sans-serif;
+body {
+    background-image: url('./assets/images/background-dark.jpg');
 }
 
-.redBtn{
-    background-color:#FF3D00;
-    border: 3px solid #DF3600;
-}
-.orangeBtn{
-    background-color: #FFB800;
-    border:3px solid #CC9300;
-    margin-top: 4%;
-}
-
-Button{
-width: 93%;
-margin:auto;
-color:#ffffff;
-font-weight: bold;
-font-size: 105%;
-height: 6vh;
-padding:6px;
-letter-spacing: -0.7px;
-}
-
-.divInfo, .divRegister{
-background-color:#FFE1BE;
-flex-direction: column;
-margin: auto;
-text-align: center;
-border-radius: 18px;
-position: absolute;
-z-index: 3;
-}
-
-.divInfo{
-    height: 60%;
-    margin-top: 10%; /*aqui se for tela cheia tem de ser 12%*/
-    width: 21vw;
-    margin-left: 37%;
-}
-.divRegister{
-    height: 50%;
-    margin-top: 12%; /*aqui se for tela cheia tem de ser 14%*/
-    width: 25vw;
-    margin-left: 35%;
-}
-Input{
-    width: 92%;
-    margin:auto;
-    border:none;
-    height: 30px;
-    outline: none;
-}
-
-.imgBurger{
+.imgBurger {
     position: absolute;
     z-index: -3;
-    margin-left: 25%;
-    height: 100vh;
-    width:45%;
+    margin-left: 26%;
+    margin-top: 6%;
+    height: 80vh;
+    width: 48%;
 }
-.burgerLogo{
-    width: 20%;
-    position:absolute;
+
+.burgerLogo {
+    width: 15%;
+    position: absolute;
     z-index: -3;
-    margin-top: 5%;
-    margin-left: 6%;
+    margin-top: 3%;
+    margin-left: 3%;
 }
-.isWorker{
-    margin-bottom: 2%;
-    margin-top: 12%;
-    color: #3B1E04;
+
+.divInfo,
+.divRegister {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    position: fixed;
+    width: 100vw;
+    height: 100vh;
 }
-.labelInputs{
+
+.formFieldsetLogin,
+.formFieldsetRegister {
+    border: none;
+    background-color: #FFE1BE;
+    margin: 3rem 0rem 0rem 0rem;
+    border-radius: 18px;
+    position: absolute;
+    z-index: 3;
+    width: 22vw;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+
+}
+
+.h1Login,
+.h1Register {
+    color: #763D09;
+    font-size: 1.5rem;
+    font-weight: 500;
+    padding: 0.5rem 0rem 0rem 0rem;
+}
+
+.labelInputs {
+    font-family: Roboto;
+    font-size: 0.8rem;
+    padding: 0rem 0rem 0rem 0.2rem;
     text-align: left;
     margin-left: 3%;
     margin-bottom: 0;
     margin-top: 3%;
     color: #3B1E04;
     letter-spacing: -0.7px;
+}
+
+.inputEmail,
+.inputPassword {
+    width: 92%;
+    margin: auto;
+    border: none;
+    height: 30px;
+    outline: none;
+    font-family: Roboto;
+}
+
+.inputPassword {
+    margin: 0rem 0rem 1rem 0rem;
+}
+
+Button {
+    cursor: pointer;
+    width: 93%;
+    color: #ffffff;
+    font-size: 1.1rem;
+    border-radius: 3px;
+    padding: 0.2rem;
+    margin: 0rem 0rem 2rem;
+}
+
+.yellowBtn {
+    background-color: #FFB800;
+    border-top: 4px solid #CC9300;
+    border-bottom: 4px solid #CC9300;
+    border-left: 4px solid #CC9300;
+    border-right: 6.5px solid #CC9300;
+}
+
+.redBtn {
+    background-color: #FF3D00;
+    border-top: 4px solid #DF3600;
+    border-bottom: 4px solid #DF3600;
+    border-left: 4px solid #DF3600;
+    border-right: 6.5px solid #DF3600;
+    font-size: 1rem;
+}
+
+.orangeBtn {
+    background-color: #FF8A00;
+    border-top: 3.5px solid #DC7700;
+    border-bottom: 3.5px solid #DC7700;
+    border-left: 3.5px solid #DC7700;
+    border-right: 6px solid #DC7700;
+}
+
+.isWorker {
+    color: #3B1E04;
+    font-family: Roboto;
+    margin-bottom: 0.3rem;
+}
+
+.backHome {
+    outline: none;
+    border: none;
+    margin: 1% 0% 38% 0%;
+    background-color: transparent;
+    position: absolute;
+    z-index: -3;
+    width: 15%;
+    font-size: 1.2rem;
+}
+
+.backHome:hover {
+    color: #FF8A00;
+    text-decoration: underline;
+    transition: 1s;
 }
