+++ src/AppResponsive.css
@@ -1,0 +1,235 @@
+@media (min-width:1400px) {
+
+    .imgBurger {
+        margin-top: 5%;
+        margin-left: 25%;
+        height: 80vh;
+        width: 50%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 25vw;
+    }
+
+    .h1Login,
+    .h1Register {
+        font-size: 1.8rem;
+    }
+
+    .labelInputs {
+        font-size: 1rem;
+    }
+
+    .inputEmail,
+    .inputPassword {
+        width: 92%;
+        height: 35px;
+    }
+
+    Button {
+        font-size: 1.3rem;
+        border-radius: 3px;
+        padding: 0.2rem;
+        margin: 0rem 0rem 2rem;
+    }
+
+    .isWorker,
+    .redBtn {
+        font-size: 1.1rem;
+    }
+
+    .backHome {
+        margin: 1% 0% 40% 0%;
+    }
+
+}
+
+@media (max-width: 1200px) {
+    .imgBurger {
+        margin-top: 8%;
+        margin-left: 22.5%;
+        height: 75vh;
+        width: 55%;
+    }
+
+    .burgerLogo {
+        width: 25%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 25vw;
+        margin: 8vh 0rem 0rem 0rem;
+    }
+
+    .backHome {
+        margin: 1% 0% 40% 0%;
+    }
+}
+
+@media (max-width: 1024px) {
+    .imgBurger {
+        margin-top: 10%;
+        width: 60vw;
+        margin-left: 20%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 25vw;
+    }
+
+    .backHome {
+        margin: 1% 0% 51% 0%;
+        font-size: 1.3rem;
+        width: 50vw;
+    }
+}
+
+@media (max-width: 764px) {
+    .imgBurger {
+        margin-top: 15%;
+        margin-left: 0%;
+        height: 70vh;
+        width: 100%;
+    }
+
+    .burgerLogo {
+        width: 28%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 50vw;
+        margin: 5vh 0rem 0rem 0rem;
+    }
+
+    .backHome {
+        width: 30vw;
+        margin: 1% 0% 60% 0%;
+    }
+}
+
+
+
+@media (max-width: 596px) {
+    .imgBurger {
+        display: none;
+    }
+
+    .burgerLogo {
+        width: 35vw;
+        margin-left: 34%;
+        margin-top: 5%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 80vw;
+        margin: 25vh 0rem 0rem 0rem;
+
+    }
+
+
+    .backHome {
+        width: 50vw;
+        margin: 1% 0% 35% 0%;
+    }
+}
+
+@media(max-width: 496px) {
+    .imgBurger {
+        display: none;
+    }
+
+    .burgerLogo {
+        width: 45vw;
+        margin-left: 27%;
+        margin-top: 8%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 80vw;
+        margin: 25vh 0rem 0rem 0rem;
+
+    }
+
+    .backHome {
+        width: 70vw;
+        margin: 1% 0% 40% 0%;
+    }
+}
+
+@media(max-width: 400px) {
+    .imgBurger {
+        display: none;
+    }
+
+    .burgerLogo {
+        width: 55vw;
+        margin-left: 22%;
+        margin-top: 10%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 80vw;
+        margin: 22vh 0rem 0rem 0rem;
+
+    }
+
+    .backHome {
+        width: 70vw;
+        margin: 1% 0% 50% 0%;
+    }
+}
+
+@media(max-width: 340px) {
+    .imgBurger {
+        display: none;
+    }
+
+    .burgerLogo {
+        width: 55vw;
+        margin-left: 22%;
+        margin-top: 10%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 80vw;
+        margin: 18vh 0rem 0rem 0rem;
+
+    }
+
+    .backHome {
+        width: 70vw;
+        margin: 1% 0% 60% 0%;
+    }
+}
+
+@media(max-width: 280px) {
+    .imgBurger {
+        display: none;
+    }
+
+    .burgerLogo {
+        width: 60vw;
+        margin-left: 20%;
+        margin-top: 18%;
+    }
+
+    .formFieldsetLogin,
+    .formFieldsetRegister {
+        width: 80vw;
+        margin: 7vh 0rem 0rem 0rem;
+
+    }
+
+    .backHome {
+        width: 70vw;
+        margin: 1% 0% 85% 0%;
+    }
+}
