+++ src/components/Button/index.js
@@ -1,13 +1,13 @@
 import React from "react";
 
 export function Button({
-btnType,
-btnClass,
-btnText,
-btnOnClick,
+    btnType,
+    btnClass,
+    btnText,
+    btnOnClick,
 
- }) {
-     return (
-         <button type={btnType} class={btnClass} onClick={btnOnClick}>{btnText}</button>
-     )
- }
+}) {
+    return (
+        <button type={btnType} class={btnClass} onClick={btnOnClick}>{btnText}</button>
+    )
+}
