+++ src/pages/Register/index.js
@@ -1,25 +1,34 @@
-import imgBurger from '../../assets/images/figma-burger.png';
-import burgerLogo from '../../assets/images/burger-logo.png';
+import { useHistory } from 'react-router-dom';
+import imgBurger from '../../assets/images/burger-background.png';
+import logo from '../../assets/images/logo.png';
 import { Button } from '../../components/Button/index.js';
 import { Input } from '../../components/Input/index.js'
 
-export function Register(){
-    return(
-<div>
-    <main>
-    <img src={burgerLogo}  className="burgerLogo" alt="logo" />
-    <img src={imgBurger} className="imgBurger" alt="imgburger" />
-    <div className="divRegister">
-                    <h1> CADASTRO</h1>
-                    <form>
-                        <p class="labelInputs">Email</p>
-                        <Input btnType="text" inputClass="inputEmail" /*inputValue={emailLogin}*/ /> 
+export function Register() {
+
+    const history = useHistory()
+
+    function navigateToLogin() {
+        history.push('/login');
+    }
+
+    return (
+        <main>
+            <img src={imgBurger} className="imgBurger" alt="imgburger" />
+            <img src={logo} className="burgerLogo" alt="logo" />
+            <div className="divRegister">
+                <Button btnType="button" btnClass="backHome" btnText="← Voltar para a Home" btnOnClick={navigateToLogin}></Button>
+                <fieldset className="formFieldsetLogin">
+                    <h1 className="h1Register"> CADASTRO</h1>
+                    <form className="formRegister">
+                        <p className="labelInputs">Email</p>
+                        <Input btnType="email" inputClass="inputEmail" /*inputValue={emailLogin}*/ />
                         <p class="labelInputs">Senha</p>
-                        <Input inputType="password" inputClass="passwordUser" /*inputValue={passwordLogin}*/ />
+                        <Input inputType="password" inputClass="inputPassword" /*inputValue={passwordLogin}*/ />
                         <Button btnType="submit" btnClass="orangeBtn" btnText="CADASTRAR" />
                     </form>
-                </div>
-    </main>
-</div>
+                </fieldset>
+            </div>
+        </main>
     )
 }
